major refactoring: change usages of Double -> double & code formatting

+++ src/main/java/com/masterisehomes/geometryapi/geodesy/SphericalMetricConversion.java
@@ -2,14 +2,20 @@ package com.masterisehomes.geometryapi.geodesy;
 
 import java.lang.Math;
 
+/* An algorithm to convert meters to degrees of Latitude & Longitude
+ * 
+ * However, this algorithm is not as good as SphericalMercatorProjection
+ * provided by OpenStreetMap wiki
+ * - https://wiki.openstreetmap.org/wiki/Mercator
+ */
 public class SphericalMetricConversion {
-    private static final double metersPerLatitude = 111111.0;
+    private static final int metersPerLatitude = 111111;
 
     public static double meterToLatitude(double meter) {
         double latitudeDegree = meter / metersPerLatitude;
         return latitudeDegree;
     }
- 
+
     public static double meterToLongitude(double meter, double latitude) {
         double longitudeDegree = meter / metersPerLatitude * Math.cos(Math.toRadians(latitude));
         return longitudeDegree;
